+++ src/Symfony/Component/Notifier/Bridge/Telegram/Tests/TelegramTransportTest.php
@@ -330,4 +330,85 @@ JSON;
 
         $transport->send(new ChatMessage('I contain special characters _ * [ ] ( ) ~ ` > # + - = | { } . ! to send.'));
     }
+
+    public function testSendPhotoWithOptions()
+    {
+        $response = $this->createMock(ResponseInterface::class);
+        $response->expects($this->exactly(2))
+            ->method('getStatusCode')
+            ->willReturn(200);
+
+        $content = <<<JSON
+            {
+                "ok": true,
+                "result": {
+                    "message_id": 1,
+                    "from": {
+                        "id": 12345678,
+                        "is_bot": true,
+                        "first_name": "YourBot",
+                        "username": "YourBot"
+                    },
+                    "chat": {
+                        "id": 1234567890,
+                        "first_name": "John",
+                        "last_name": "Doe",
+                        "username": "JohnDoe",
+                        "type": "private"
+                    },
+                    "date": 1459958199,
+                    "photo": [
+                        {
+                            "file_id": "ABCDEF",
+                            "file_unique_id" : "ABCDEF1",
+                            "file_size": 1378,
+                            "width": 90,
+                            "height": 51
+                        },
+                        {
+                            "file_id": "ABCDEF",
+                            "file_unique_id" : "ABCDEF2",
+                            "file_size": 19987,
+                            "width": 320,
+                            "height": 180
+                        }
+                    ],
+                    "caption": "Hello from Bot!"
+                }
+            }
+JSON;
+
+        $response->expects($this->once())
+            ->method('getContent')
+            ->willReturn($content)
+        ;
+
+        $expectedBody = [
+            'photo' => 'https://image.ur.l/',
+            'has_spoiler' => true,
+            'chat_id' => 'testChannel',
+            'parse_mode' => 'MarkdownV2',
+            'caption' => 'testMessage',
+        ];
+
+        $client = new MockHttpClient(function (string $method, string $url, array $options = []) use ($response, $expectedBody): ResponseInterface {
+            $this->assertStringEndsWith('/sendPhoto', $url);
+            $this->assertSame($expectedBody, json_decode($options['body'], true));
+
+            return $response;
+        });
+
+        $transport = self::createTransport($client, 'testChannel');
+
+        $messageOptions = new TelegramOptions();
+        $messageOptions
+            ->photo('https://image.ur.l/')
+            ->hasSpoiler(true)
+        ;
+
+        $sentMessage = $transport->send(new ChatMessage('testMessage', $messageOptions));
+
+        $this->assertEquals(1, $sentMessage->getMessageId());
+        $this->assertEquals('telegram://api.telegram.org?channel=testChannel', $sentMessage->getTransport());
+    }
 }